feat: palettes and default theme updates
