+++ packages/ui/src/theme/defaults/dark-theme.ts
@@ -1,100 +1,36 @@
 import {
-  grayDark,
-  mauveDark,
   slateDark,
-  sageDark,
-  oliveDark,
-  sandDark,
-  tomatoDark,
-  redDark,
+  whiteA,
   crimsonDark,
-  pinkDark,
-  plumDark,
-  purpleDark,
-  violetDark,
-  indigoDark,
-  blueDark,
-  skyDark,
-  mintDark,
-  cyanDark,
-  tealDark,
-  greenDark,
-  grassDark,
-  limeDark,
-  yellowDark,
   amberDark,
-  orangeDark,
-  brownDark,
-  bronzeDark,
-  goldDark,
-  grayDarkA,
-  mauveDarkA,
-  slateDarkA,
-  sageDarkA,
-  oliveDarkA,
-  sandDarkA,
-  tomatoDarkA,
-  redDarkA,
-  crimsonDarkA,
-  pinkDarkA,
-  plumDarkA,
-  purpleDarkA,
-  violetDarkA,
-  indigoDarkA,
-  blueDarkA,
-  skyDarkA,
-  mintDarkA,
-  cyanDarkA,
-  tealDarkA,
-  greenDarkA,
-  grassDarkA,
-  limeDarkA,
-  yellowDarkA,
-  amberDarkA,
-  orangeDarkA,
-  brownDarkA,
-  bronzeDarkA,
-  goldDarkA,
+  tealDark,
+  indigoDark,
+  purpleDark,
+  blueDark,
 } from '@radix-ui/colors'
 
-export const darkColors = {
-  ...grayDark,
-  ...mauveDark,
-  ...slateDark,
-  ...sageDark,
-  ...oliveDark,
-  ...sandDark,
-  ...tomatoDark,
-  ...redDark,
-  ...crimsonDark,
-  ...pinkDark,
-  ...plumDark,
-  ...purpleDark,
-  ...violetDark,
-  ...indigoDark,
-  ...blueDark,
-  ...skyDark,
-  ...mintDark,
-  ...cyanDark,
-  ...tealDark,
-  ...greenDark,
-  ...grassDark,
-  ...limeDark,
-  ...yellowDark,
-  ...amberDark,
-  ...orangeDark,
-  ...brownDark,
-  ...bronzeDark,
-  ...goldDark,
-
-  // Semantic colors
-  hiContrast: '$slate12',
-  loContrast: '$slate1',
-  canvas: 'hsl(0 0% 15%)',
-  panel: '$slate3',
-  transparentPanel: 'hsl(0 100% 100% / 97%)',
-  shadowLight: 'hsl(206 22% 7% / 35%)',
-  shadowDark: 'hsl(206 22% 7% / 20%)',
+export const colors = {
+  hiContrast: slateDark.slate12,
+  loContrast: whiteA.whiteA12,
+  error: crimsonDark.crimson7,
+  warning: amberDark.amber7,
+  success: tealDark.teal9,
+  info: indigoDark.indigo9,
+  panel: 'hsl(0 0% 93%)',
+  // branding
+  primary: purpleDark.purple7,
+  accent: tealDark.teal7,
+  secondary: indigoDark.indigo7,
+  highlight: blueDark.blue7,
+  muted: slateDark.slate7,
+  // // Semantic colors
+  // hiContrast: '$slate12',
+  // loContrast: '$slate1',
+  // canvas: 'hsl(0 0% 15%)',
+  // panel: '$slate3',
+  // transparentPanel: 'hsl(0 100% 100% / 97%)',
+  // shadowLight: 'hsl(206 22% 7% / 35%)',
+  // shadowDark: 'hsl(206 22% 7% / 20%)',
 }
 
 /**
@@ -126,4 +62,4 @@ export const darkColors = {
  *
  * 12: High-contrast text
  */
-export type AvailableDarkColors = keyof typeof darkColors
+export type AvailableDarkColors = keyof typeof colors
