+++ packages/ui/src/theme/defaults/index.ts
@@ -1,5 +1,6 @@
 export type { AvailableLightColors } from './light-theme'
 export type { AvailableDarkColors } from './dark-theme'
-export { lightColors } from './light-theme'
-export { darkColors } from './dark-theme'
-export { theme } from './common'
+export { colors as lightColors } from './light-theme'
+export { colors as darkColors } from './dark-theme'
+export * from './tokens'
+export * from './utils'
