+++ packages/ui/src/theme/defaults/palettes/hero.ts
@@ -1,0 +1,7 @@
+export const hero = {
+  secondary: '#ec3c48',
+  primary: '#0066ff',
+  highlight: '#ff8a05',
+  tertiary: '#4E0096',
+  white: '#f7f7f7',
+}
