+++ packages/ui/src/theme/defaults/palettes/index.ts
@@ -1,0 +1,2 @@
+export * from './hero'
+export * from './plum'
