+++ packages/ui/src/theme/defaults/palettes/plum.ts
@@ -1,0 +1,7 @@
+export const plum = {
+  secondary: '#05181a',
+  primary: '#541cb4',
+  highlight: '#ea1278',
+  tertiary: '#fdb0cd',
+  white: '#eff3f1',
+}
