+++ packages/ui/src/theme/defaults/palettes/trubit.json
@@ -1,0 +1,47 @@
+{
+  "color-primary-100": "#E0D1FE",
+  "color-primary-200": "#BFA4FE",
+  "color-primary-300": "#9C77FC",
+  "color-primary-400": "#7F55F9",
+  "color-primary-500": "#501EF6",
+  "color-primary-600": "#3D15D3",
+  "color-primary-700": "#2C0FB1",
+  "color-primary-800": "#1E098E",
+  "color-primary-900": "#140576",
+  "color-success-100": "#F1FDCF",
+  "color-success-200": "#DFFB9F",
+  "color-success-300": "#C6F56E",
+  "color-success-400": "#ACEB49",
+  "color-success-500": "#86de13",
+  "color-success-600": "#6ABE0D",
+  "color-success-700": "#519F09",
+  "color-success-800": "#3B8006",
+  "color-success-900": "#2B6A03",
+  "color-info-100": "#CCFFFA",
+  "color-info-200": "#99FFFD",
+  "color-info-300": "#66F6FF",
+  "color-info-400": "#3FE6FF",
+  "color-info-500": "#00CBFF",
+  "color-info-600": "#009EDB",
+  "color-info-700": "#0076B7",
+  "color-info-800": "#005493",
+  "color-info-900": "#003D7A",
+  "color-warning-100": "#FFF4CC",
+  "color-warning-200": "#FFE799",
+  "color-warning-300": "#FFD666",
+  "color-warning-400": "#FFC53F",
+  "color-warning-500": "#FFAA00",
+  "color-warning-600": "#DB8A00",
+  "color-warning-700": "#B76D00",
+  "color-warning-800": "#935300",
+  "color-warning-900": "#7A4000",
+  "color-danger-100": "#FFE4D3",
+  "color-danger-200": "#FFC3A8",
+  "color-danger-300": "#FF9A7C",
+  "color-danger-400": "#FF745C",
+  "color-danger-500": "#FF3426",
+  "color-danger-600": "#DB1B1F",
+  "color-danger-700": "#B71323",
+  "color-danger-800": "#930C25",
+  "color-danger-900": "#7A0726"
+}
